Add generalized cache handling

+++ src/SettingsResult.php
@@ -9,10 +9,10 @@ class SettingsResult
 {
     /** @var array */
     public $settings;
-    /** @var int */
+    /** @var float */
     public $fetchTime;
 
-    public function __construct(?array $settings, int $fetchTime)
+    public function __construct(?array $settings, float $fetchTime)
     {
         $this->settings = $settings;
         $this->fetchTime = $fetchTime;
